[IDLE-241] refactor: 패키지 구조 개선
 - 애플 소셜로그인 콜백 api를 AuthenticationController로 옮기고 기존 AppleController 삭제
 - DTO명 변경: AppleDTO -> AppleOAuth2TokenInfo, AppleUser -> AppleOAuth2UserInfo
 - AppleUtil 메서드 AppleOAuth2UserService 내에 프라이빗 메서드로 이동
 - 서비스명 변경: AppleService -> AppleOAuth2UserService
 - 애플 소셜로그인은 스프링 시큐리티를 활용하지 않기 때문에 OAuthAttributes에서 애플 사용자 처리 로직 삭제
