--- user-service/src/main/java/kr/mybrary/userservice/authentication/domain/oauth2/apple/AppleUser.java
@@ -1,44 +1,0 @@
-package kr.mybrary.userservice.authentication.domain.oauth2.apple;
-
-public class AppleUser {
-    private Name name;
-    private String email;
-
-    public Name getName() {
-        return name;
-    }
-
-    public void setName(Name name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    static class Name {
-        private String firstName;
-        private String lastName;
-
-        public String getFirstName() {
-            return firstName;
-        }
-
-        public void setFirstName(String firstName) {
-            this.firstName = firstName;
-        }
-
-        public String getLastName() {
-            return lastName;
-        }
-
-        public void setLastName(String lastName) {
-            this.lastName = lastName;
-        }
-    }
-
-}
