+++ user-service/src/main/java/kr/mybrary/userservice/authentication/domain/oauth2/service/AppleOAuth2UserService.java
@@ -1,9 +1,16 @@
-package kr.mybrary.userservice.authentication.domain.oauth2.apple;
+package kr.mybrary.userservice.authentication.domain.oauth2.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nimbusds.jose.JOSEException;
+import com.nimbusds.jose.JWSAlgorithm;
+import com.nimbusds.jose.JWSHeader;
+import com.nimbusds.jose.JWSSigner;
+import com.nimbusds.jose.crypto.ECDSASigner;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import jakarta.servlet.http.HttpServletRequest;
+import kr.mybrary.userservice.authentication.domain.oauth2.userinfo.AppleOAuth2TokenInfo;
+import kr.mybrary.userservice.authentication.domain.oauth2.userinfo.AppleOAuth2UserInfo;
 import kr.mybrary.userservice.global.util.JwtUtil;
 import kr.mybrary.userservice.global.util.RedisUtil;
 import kr.mybrary.userservice.user.persistence.Role;
@@ -15,6 +22,10 @@ import lombok.extern.slf4j.Slf4j;
 import net.minidev.json.JSONObject;
 import net.minidev.json.parser.JSONParser;
 import org.apache.commons.lang3.RandomStringUtils;
+import org.bouncycastle.util.io.pem.PemObject;
+import org.bouncycastle.util.io.pem.PemReader;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
@@ -25,10 +36,16 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
+import java.io.FileReader;
+import java.io.IOException;
+import java.security.KeyFactory;
+import java.security.NoSuchAlgorithmException;
+import java.security.interfaces.ECPrivateKey;
+import java.security.spec.InvalidKeySpecException;
+import java.security.spec.PKCS8EncodedKeySpec;
 import java.time.Duration;
 import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Date;
 import java.util.UUID;
 
 import static kr.mybrary.userservice.global.constant.ImageConstant.*;
@@ -36,7 +53,7 @@ import static kr.mybrary.userservice.global.constant.ImageConstant.*;
 @RequiredArgsConstructor
 @Service
 @Slf4j
-public class AppleService {
+public class AppleOAuth2UserService {
 
     private String APPLE_TEAM_ID = "KU782MK78R";
 
@@ -44,7 +61,7 @@ public class AppleService {
 
     private String APPLE_CLIENT_ID = "kr.mybrary.service";
 
-    private String APPLE_REDIRECT_URL = "https://8a2c-118-33-100-24.ngrok-free.app/apple/callback";
+    private String APPLE_REDIRECT_URL = "https://f97d-1-235-157-76.ngrok-free.app/auth/v1/apple/callback";
 
     private String APPLE_KEY_PATH = "apple/AuthKey_8LCJC23RYD.p8";
     static final String CALLBACK_URL = "kr.mybrary://";
@@ -53,7 +70,6 @@ public class AppleService {
     static final int REFRESH_TOKEN_EXPIRATION = 14;
 
     private final static String APPLE_AUTH_URL = "https://appleid.apple.com";
-    private final AppleUtil appleUtil;
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
     private final JwtUtil jwtUtil;
@@ -62,12 +78,12 @@ public class AppleService {
     public String getAppleInfo(HttpServletRequest request) throws Exception {
         log.info("Apple Login Start");
         ObjectMapper objectMapper = new ObjectMapper();
-        AppleUser appleUser;
+        AppleOAuth2UserInfo appleUser;
         String fullName = "";
 
         // user 정보가 있으면 AppleUser 객체로 변환
         if(request.getParameter("user") != null) {
-            appleUser = objectMapper.readValue(request.getParameter("user"), AppleUser.class);
+            appleUser = objectMapper.readValue(request.getParameter("user"), AppleOAuth2UserInfo.class);
             fullName = appleUser.getName().getLastName() + appleUser.getName().getFirstName();
 //            String name = appleUser.getName().getLastName() + appleUser.getName().getFirstName();
 //            String email = appleUser.getEmail();
@@ -77,7 +93,7 @@ public class AppleService {
         if (request.getParameter("code") == null) throw new Exception("Failed get authorization code");
 
         // Client Secret 생성
-        String clientSecret = appleUtil.createAppleClientSecret(APPLE_CLIENT_ID, "AuthKey_8LCJC23RYD.p8/8LCJC23RYD/KU782MK78R");
+        String clientSecret = createAppleClientSecret(APPLE_CLIENT_ID, "AuthKey_8LCJC23RYD.p8/8LCJC23RYD/KU782MK78R");
         String userId = "";
         String email  = "";
         String accessToken = "";
@@ -94,6 +110,7 @@ public class AppleService {
             params.add("client_secret", clientSecret);
             params.add("code"         , request.getParameter("code"));
             params.add("redirect_uri" , APPLE_REDIRECT_URL);
+
 
             RestTemplate restTemplate = new RestTemplate(); // TODO: FeignClient로 변경
             HttpEntity<MultiValueMap<String, String>> httpEntity = new HttpEntity<>(params, headers);
@@ -125,7 +142,7 @@ public class AppleService {
         }
 
         // 가입된 회원인지 확인
-        AppleDTO appleDTO = AppleDTO.builder()
+        AppleOAuth2TokenInfo appleOAuth2TokenInfo = AppleOAuth2TokenInfo.builder()
                 .id(userId)
                 .token(accessToken)
                 .email(email)
@@ -133,7 +150,7 @@ public class AppleService {
                 .build();
 
 
-        User createdUser = getUser(appleDTO);
+        User createdUser = getUser(appleOAuth2TokenInfo);
 
         String my_accessToken = jwtUtil.createAccessToken(createdUser.getLoginId(), LocalDateTime.now());
         String my_refreshToken = jwtUtil.createRefreshToken(LocalDateTime.now());
@@ -150,7 +167,7 @@ public class AppleService {
     }
 
     public void withdrawApple(String socialId) {
-        String clientSecret = appleUtil.createAppleClientSecret(APPLE_CLIENT_ID, "AuthKey_8LCJC23RYD.p8/8LCJC23RYD/KU782MK78R");
+        String clientSecret = createAppleClientSecret(APPLE_CLIENT_ID, "AuthKey_8LCJC23RYD.p8/8LCJC23RYD/KU782MK78R");
         String refreshToken = (String) redisUtil.get("APPLE_"+socialId);
 
         HttpHeaders headers = new HttpHeaders();
@@ -178,23 +195,23 @@ public class AppleService {
         redisUtil.delete("APPLE_"+socialId);
     }
 
-    private User getUser(AppleDTO appleDTO) {
-        User findUser = userRepository.findBySocialTypeAndSocialId(SocialType.APPLE, appleDTO.getId()).orElse(null);
+    private User getUser(AppleOAuth2TokenInfo appleOAuth2TokenInfo) {
+        User findUser = userRepository.findBySocialTypeAndSocialId(SocialType.APPLE, appleOAuth2TokenInfo.getId()).orElse(null);
 
         if (findUser == null) {
-            return saveUser(appleDTO);
+            return saveUser(appleOAuth2TokenInfo);
         }
         return findUser;
     }
 
-    private User saveUser(AppleDTO appleDTO) {
+    private User saveUser(AppleOAuth2TokenInfo appleOAuth2TokenInfo) {
         User createdUser = User.builder()
                 .socialType(SocialType.APPLE)
-                .socialId(appleDTO.getId())
+                .socialId(appleOAuth2TokenInfo.getId())
                 .loginId(UUID.randomUUID().toString())
                 .password(UUID.randomUUID().toString())
-                .email(appleDTO.getEmail())
-                .nickname(appleDTO.getFullName() + RandomStringUtils.randomNumeric(5))
+                .email(appleOAuth2TokenInfo.getEmail())
+                .nickname(appleOAuth2TokenInfo.getFullName() + RandomStringUtils.randomNumeric(5))
                 .introduction("")
                 .role(Role.USER)
                 .build();
@@ -209,5 +226,53 @@ public class AppleService {
         createdUser.updateProfileImageThumbnailSmallUrl(DEFAULT_PROFILE_IMAGE_SMALL.getUrl());
     }
 
+    private String createAppleClientSecret(String clientId, String clientSecret) {
+        String[] secretKeyResourceArr = clientSecret.split("/");
+        String appleKeyId = secretKeyResourceArr[1];
+        String appleTeamId = secretKeyResourceArr[2];
+        String appleClientId = clientId;
+
+        JWSHeader header = new JWSHeader.Builder(JWSAlgorithm.ES256).keyID(appleKeyId).build();
+        JWTClaimsSet claimsSet = new JWTClaimsSet.Builder()
+                .issuer(appleTeamId)
+                .subject(appleClientId)
+                .issueTime(new Date())
+                .expirationTime(new Date(System.currentTimeMillis() + 1000 * 60 * 5))
+                .audience("https://appleid.apple.com")
+                .build();
+
+        SignedJWT jwt = new SignedJWT(header, claimsSet);
+
+        PKCS8EncodedKeySpec spec = new PKCS8EncodedKeySpec(readPrivateKey("/apple/AuthKey_8LCJC23RYD.p8"));
+        KeyFactory kf;
+        try {
+            kf = KeyFactory.getInstance("EC");
+        } catch (NoSuchAlgorithmException e) {
+            throw new RuntimeException(e);
+        }
+
+        try {
+            ECPrivateKey privateKey = (ECPrivateKey) kf.generatePrivate(spec);
+            JWSSigner jwsSigner = new ECDSASigner(privateKey);
+            jwt.sign(jwsSigner);
+        } catch (InvalidKeySpecException e) {
+            throw new RuntimeException(e);
+        } catch (JOSEException e) {
+            throw new RuntimeException(e);
+        }
+        return jwt.serialize();
+    }
+
+    private byte[] readPrivateKey(String keyPath) {
+        Resource resource = new ClassPathResource(keyPath);
+        try {
+            FileReader keyReader = new FileReader(resource.getFile());
+            PemReader pemReader = new PemReader(keyReader);
+            PemObject pemObject = pemReader.readPemObject();
+            return pemObject.getContent();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
 
 }
