+++ user-service/src/main/java/kr/mybrary/userservice/authentication/domain/oauth2/userinfo/AppleOAuth2TokenInfo.java
@@ -1,11 +1,11 @@
-package kr.mybrary.userservice.authentication.domain.oauth2.apple;
+package kr.mybrary.userservice.authentication.domain.oauth2.userinfo;
 
 import lombok.Builder;
 import lombok.Getter;
 
 @Builder
 @Getter
-public class AppleDTO {
+public class AppleOAuth2TokenInfo {
 
     private String id;
     private String token;
