+++ user-service/src/main/java/kr/mybrary/userservice/authentication/presentation/AuthenticationController.java
@@ -3,11 +3,13 @@ package kr.mybrary.userservice.authentication.presentation;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import kr.mybrary.userservice.authentication.domain.AuthenticationService;
+import kr.mybrary.userservice.authentication.domain.oauth2.service.AppleOAuth2UserService;
 import kr.mybrary.userservice.global.dto.response.SuccessResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthenticationController {
 
     private final AuthenticationService authenticationService;
+    private final AppleOAuth2UserService appleOAuth2UserService;
 
     @GetMapping("/refresh")
     public ResponseEntity<SuccessResponse<Void>> refreshToken(HttpServletRequest request, HttpServletResponse response) {
@@ -25,6 +28,11 @@ public class AuthenticationController {
         return ResponseEntity.status(200).body(
                 SuccessResponse.of(HttpStatus.CREATED.toString(), "토큰 재발급에 성공했습니다.", null)
         );
+    }
+
+    @PostMapping("/apple/callback")
+    public void callback(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        response.sendRedirect(appleOAuth2UserService.getAppleInfo(request));
     }
 
 }
